% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goproxy.R
\docType{package}
\name{goproxy-package}
\alias{goproxy}
\alias{goproxy-package}
\title{goproxy: Reverse proxy in Go}
\description{
Starts a simple and small reverse proxy on localhost. The reverse proxy is implemented in golang
}
\author{
\strong{Maintainer}: Cole Arendt \email{cole@rstudio.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
