% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_proxy.R
\name{run_proxy}
\alias{run_proxy}
\title{Run a reverse proxy}
\usage{
run_proxy(url, port = "9845", timeout = 20)
}
\arguments{
\item{url}{The url (with protocol) to reverse proxy}

\item{port}{The port on which to listen}
}
\description{
Starts a reverse proxy in a subprocess (for process cleanup),
but monitors the process for stdout / stderr output.
}
