% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_proxy.R
\name{run_proxy_raw}
\alias{run_proxy_raw}
\title{Run a reverse proxy directly}
\usage{
run_proxy_raw(url, port = "9845")
}
\arguments{
\item{url}{The url (with protocol) to proxy}

\item{port}{The port on which to listen}
}
\description{
WARNING: this is currently not able to be interrupted. Use `run_proxy`
instead.
}
